refactor: simplify TaskDetailsDialogActionButtons by using custom hook
- Replaced inline logic with a custom hook `useTaskDetailsDialogActionButtons` to manage state and actions for task details.
- Removed unnecessary props and improved button disable logic based on editing permissions.
- Streamlined the component structure for better readability and maintainability.
